Splits BigQueryIOTest into a read and a write test
